Move checked allocation functions to their own translation unit
